move IsValidTid to common
